+++ packages/board-arduino-uno/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @typecad/board-arduino-uno
 
+## 0.1.0-alpha.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @typecad/cuttlefish@0.1.0-alpha.2
+  - @typecad/mcu-atmega328p@0.1.0-alpha.2
+  - @typecad/hal@0.1.0-alpha.2
+
 ## 0.1.0-alpha.1
 
 ### Minor Changes