+++ packages/hal/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @typecad/hal
 
+## 0.1.0-alpha.2
+
 ## 0.1.0-alpha.1
 
 ### Minor Changes