+++ packages/mcu-esp32/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @typecad/mcu-esp32
 
+## 0.1.0-alpha.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @typecad/cuttlefish@0.1.0-alpha.2
+  - @typecad/hal@0.1.0-alpha.2
+
 ## 0.1.0-alpha.1
 
 ### Minor Changes