+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @typecad/ui
 
+## 0.1.0-alpha.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @typecad/cuttlefish@0.1.0-alpha.2
+
 ## 0.1.0-alpha.1
 
 ### Minor Changes